web_scraper/requirements.txt
fastapi
psycopg2-binary
sqlalchemy
pydantic
requests
celery
redis
uvicorn
alembic
python-dotenv
python-decouple